Update sensitive cmd install instructions (#43)
On 64-bit machines, the previous installation instructions to monitor process executions would miss 32-bit executions, as this needs different registry keys set. By explicitly installing the registery keys for both 32 and 64-bit this will ensure the alert always gets triggered.

We considered specifying the WOW6432Node version of the registry keys, but that only worked when installed via the 64-bit view of registry. If installed via the 32-bit view, the 64-bit executions would go unmonitored as the un-adorned register key would be redirected to the 32-bit WOW6432Node key (which would just be written twice.)